Move monitoring.utils to _internal_utils.arg_handler
