+++ client/verta/verta/monitoring/alert/_alerter.py
@@ -5,8 +5,8 @@ import abc
 from verta.external import six
 
 from verta._protos.public.monitoring import Alert_pb2 as _AlertService
+from verta._internal_utils import arg_handler
 from verta.monitoring import comparison as comparison_module
-from verta.monitoring import utils
 
 
 # TODO: move into separate files
@@ -165,7 +165,7 @@ class ReferenceAlerter(_Alerter):
 
     def __init__(self, comparison, reference_sample):
         self._comparison = _Alerter._validate_comparison(comparison)
-        self._reference_sample_id = utils.extract_id(reference_sample)
+        self._reference_sample_id = arg_handler.extract_id(reference_sample)
 
     @property
     def comparison(self):
