+++ client/verta/verta/monitoring/labels.py
@@ -8,8 +8,7 @@ from verta._protos.public.monitoring.Labels_pb2 import (
     FindSampleLabelValuesRequest,
 )
 from .summaries.queries import _labels_proto
-from verta._internal_utils import time_utils
-from .utils import maybe
+from verta._internal_utils import arg_handler, time_utils
 
 
 class Labels(object):
@@ -127,20 +126,20 @@ class Labels(object):
 
     def _build_summary_filter(self, **kwargs):
         summary_query = kwargs.get("summary_query", None)
-        summaries_proto = maybe(lambda q: q._to_proto_request(), summary_query)
+        summaries_proto = arg_handler.maybe(lambda q: q._to_proto_request(), summary_query)
 
         sample_ids = kwargs.get("sample_id", None)
 
         window_start = kwargs.get("time_window_start", None)
-        window_start_at_millis = maybe(
+        window_start_at_millis = arg_handler.maybe(
             lambda t: time_utils.epoch_millis(t), window_start
         )
 
         window_end = kwargs.get("time_window_end", None)
-        window_end_at_millis = maybe(lambda t: time_utils.epoch_millis(t), window_end)
+        window_end_at_millis = arg_handler.maybe(lambda t: time_utils.epoch_millis(t), window_end)
 
         labels = kwargs.get("labels", None)
-        labels_proto = maybe(_labels_proto, labels)
+        labels_proto = arg_handler.maybe(_labels_proto, labels)
         return FilterQuerySummarySample(
             find_summaries=summaries_proto,
             sample_ids=sample_ids,
